+++ input.h
@@ -1,7 +1,7 @@
 /*
  * Input Buffer
  *
- * Copyright (c) 2016-2020 Alexei A. Smekalkine <ikle@ikle.ru>
+ * Copyright (c) 2016-2021 Alexei A. Smekalkine <ikle@ikle.ru>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
@@ -23,7 +23,7 @@ struct input {
 int  input_init (struct input *o, size_t size, input_reader *read, void *ctx);
 void input_fini (struct input *o);
 
-int input_fill (struct input *o);
+size_t input_fill (struct input *o);
 int input_eat  (struct input *o, size_t count);
 
 #endif  /* INPUT_H */
